+++ src/overview/onboarding/utils.ts
@@ -1,4 +1,4 @@
-import { STORAGE_KEYS } from './constants'
+import { STORAGE_KEYS, ANNOTATION_DEMO_URL } from './constants'
 import { getLocalStorage, setLocalStorage } from 'src/util/storage'
 
 export const conditionallySkipToTimeFilter = async () => {
@@ -23,3 +23,14 @@ export const conditionallySkipToTimeFilter = async () => {
  */
 export const isDuringInstall = (): boolean =>
     location.href.indexOf('install=true') > -1
+
+// Finding the coordinates to center the notification box
+export const getPageCenter = () => ({
+    x: window.innerWidth / 2,
+    y: window.innerHeight / 2 - 200, // Assuming average height as 200, since height is 'auto'
+})
+
+/**
+ * Check whether page is onboarding demo page (Wikipedia Memex)
+ */
+export const isDemoPage = () => window.location.href === ANNOTATION_DEMO_URL
